requirements.txt
streamlit
requests
beautifulsoup4
pandas
python-dotenv
greenapi_client
twilio
